--- .eslintrc.js
@@ -1,32 +1,0 @@
-module.exports = {
-  root: true,
-  env: {
-    browser: true,
-    amd: true,
-    node: true,
-    es6: true,
-  },
-  extends: [
-    'eslint:recommended',
-    'plugin:jsx-a11y/recommended',
-    'plugin:prettier/recommended',
-    'next',
-    'next/core-web-vitals',
-  ],
-  globals: {
-    __PATH_PREFIX__: true,
-  },
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    ecmaFeatures: {
-      jsx: true,
-    },
-    project: './tsconfig.json',
-  },
-  rules: {
-    'prettier/prettier': 'error',
-    'react/prop-types': 0,
-    'no-unused-vars': 0,
-    'react/no-unescaped-entities': 0,
-  },
-}
