+++ .eslintrc.json
@@ -1,0 +1,6 @@
+{
+  "extends": ["eslint:recommended", "next", "next/core-web-vitals", "prettier"],
+  "rules": {
+    "no-unused-vars": 1
+  }
+}
