+++ components/molecules/Pre/index.ts
@@ -1,0 +1,1 @@
+export { Pre } from './Pre'
