+++ components/templates/common/style.ts
