+++ package.json
@@ -49,8 +49,6 @@
     "eslint": "8.23.0",
     "eslint-config-next": "12.2.5",
     "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-jsx-a11y": "6.6.1",
-    "eslint-plugin-prettier": "4.2.1",
     "file-loader": "6.2.0",
     "husky": "8.0.1",
     "lint-staged": "13.0.3",
